Added feature to blacklist jwt tokens

+++ Online Shopping System/src/main/java/com/tothenew/repos/SellerRepository.java
@@ -3,5 +3,5 @@ package com.tothenew.repos;
 import com.tothenew.entities.user.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerRepository extends JpaRepository<Seller, Integer> {
+public interface SellerRepository extends JpaRepository<Seller, Long> {
 }
